+++ source/extensions/filters/common/local_ratelimit/local_ratelimit_impl.h
@@ -13,6 +13,7 @@
 #include "source/common/common/thread_synchronizer.h"
 #include "source/common/common/token_bucket_impl.h"
 #include "source/common/protobuf/protobuf.h"
+#include "source/extensions/filters/common/local_ratelimit/dynamic_descriptor.h"
 #include "source/extensions/filters/common/local_ratelimit/local_ratelimit.h"
 
 namespace Envoy {
@@ -22,49 +23,7 @@ namespace Common {
 namespace LocalRateLimit {
 
 class LocalRateLimiterImpl;
-class RateLimitTokenBucket;
-using RateLimitTokenBucketSharedPtr = std::shared_ptr<RateLimitTokenBucket>;
 using ProtoLocalClusterRateLimit = envoy::extensions::common::ratelimit::v3::LocalClusterRateLimit;
-
-class DynamicDescriptor : public Logger::Loggable<Logger::Id::rate_limit_quota> {
-public:
-  DynamicDescriptor(uint64_t max_tokens, uint64_t tokens_per_fill,
-                    std::chrono::milliseconds fill_interval, uint32_t lru_size,
-                    TimeSource& time_source, bool shadow_mode);
-  // add a new user configured descriptor to the set.
-  RateLimitTokenBucketSharedPtr addOrGetDescriptor(const RateLimit::Descriptor& request_descriptor);
-
-private:
-  using LruList = std::list<RateLimit::Descriptor>;
-
-  mutable absl::Mutex dyn_desc_lock_;
-  RateLimit::Descriptor::Map<std::pair<RateLimitTokenBucketSharedPtr, LruList::iterator>>
-      dynamic_descriptors_ ABSL_GUARDED_BY(dyn_desc_lock_);
-
-  uint64_t max_tokens_;
-  uint64_t tokens_per_fill_;
-  const std::chrono::milliseconds fill_interval_;
-  LruList lru_list_;
-  uint32_t lru_size_;
-  TimeSource& time_source_;
-  const bool shadow_mode_{false};
-};
-
-using DynamicDescriptorSharedPtr = std::shared_ptr<DynamicDescriptor>;
-
-class DynamicDescriptorMap : public Logger::Loggable<Logger::Id::rate_limit_quota> {
-public:
-  // add a new user configured descriptor to the set.
-  void addDescriptor(const RateLimit::LocalDescriptor& descriptor,
-                     DynamicDescriptorSharedPtr dynamic_descriptor);
-  // pass request_descriptors to the dynamic descriptor set to get the token bucket.
-  RateLimitTokenBucketSharedPtr getBucket(const RateLimit::Descriptor);
-
-private:
-  bool matchDescriptorEntries(const std::vector<RateLimit::DescriptorEntry>& request_entries,
-                              const std::vector<RateLimit::DescriptorEntry>& user_entries);
-  RateLimit::LocalDescriptor::Map<DynamicDescriptorSharedPtr> config_descriptors_;
-};
 
 class ShareProvider {
 public:
@@ -143,13 +102,13 @@ public:
   ~LocalRateLimiterImpl() override;
 
   LocalRateLimiter::Result
-  requestAllowed(absl::Span<const RateLimit::Descriptor> request_descriptors) override;
+  requestAllowed(absl::Span<const Envoy::RateLimit::Descriptor> request_descriptors) override;
 
 private:
   RateLimitTokenBucketSharedPtr default_token_bucket_;
 
   TimeSource& time_source_;
-  RateLimit::LocalDescriptor::Map<RateLimitTokenBucketSharedPtr> descriptors_;
+  Envoy::RateLimit::LocalDescriptor::Map<RateLimitTokenBucketSharedPtr> descriptors_;
 
   DynamicDescriptorMap dynamic_descriptors_{};
   ShareProviderSharedPtr share_provider_;
@@ -160,7 +119,7 @@ private:
 
 class AlwaysDenyLocalRateLimiter : public LocalRateLimiter {
 public:
-  LocalRateLimiter::Result requestAllowed(absl::Span<const RateLimit::Descriptor>) override {
+  LocalRateLimiter::Result requestAllowed(absl::Span<const Envoy::RateLimit::Descriptor>) override {
     return {false, nullptr};
   }
 };
